+++ priv/repo/migrations/20201118201429_add_category_to_published_message_projection.exs
@@ -1,0 +1,9 @@
+defmodule DealogBackoffice.Repo.Migrations.AddCategoryToPublishedMessageProjection do
+  use Ecto.Migration
+
+  def change do
+    alter table(:published_messages) do
+      add :category, :string
+    end
+  end
+end